refactor: update import paths for game schema in update actions and edit game component
